Refactored translator tests and moved them into their own test folder.
